requirements.txt
amqp==2.6.1
aniso8601==7.0.0
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
black==21.10b0
brotli==1.0.9
celery==4.4.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
cloudinary==1.24.0
cryptography==35.0.0
cssselect2==0.4.1
defusedxml==0.7.1
diff-match-patch==20200713
django-allauth==0.45.0
django-celery-beat==2.2.0
django-celery-results==2.0.1
django-cors-headers==3.10.0
django-crispy-forms==1.11.0
django-embed-video==1.4.0
django-import-export==2.6.1
django-jet==1.0.8
django-redis==5.0.0
django-summernote==0.8.11.6
django-timezone-field==4.2.3; python_version >= '3.5'
djangorestframework==3.13.1
django==3.2
docopt==0.4.0
et-xmlfile==1.1.0
flake8==4.0.1
flower==0.9.3
fonttools==4.28.1
graphene-django==2.15.0
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
html5lib==1.1
idna==3.3
iniconfig==1.1.1
iso8601==1.0.0
kombu==4.6.11

mccabe==0.6.1
mypy-extensions==0.4.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
openai==0.27.2
packaging==21.3; python_version >= '3.6'
pathspec==0.9.0
pillow==9
platformdirs==2.4.0
pluggy==1.0.0
promise==2.3

py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydyf==0.1.2
pyflakes==2.4.0
pyjwt==2.4.0
pyparsing==2.4.7
pyphen==0.11.0
pytest-django==4.4.0
pytest==6.2.5
python-crontab==2.6.0
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python3-openid==3.2.0
pytz==2021.3
pyyaml==6.0
redis==3.5.3
regex==2021.11.10
requests-oauthlib==1.3.0
requests==2.26.0
rx==1.6.1
sentry-sdk==1.4.3
setuptools==62.3.2; python_version >= '3.7'
singledispatch==3.7.0
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.1.0
sqlparse==0.4.2
tablib==3.1.0
text-unidecode==1.3
tinycss2==1.1.0
toml==0.10.2
tomli==1.2.2
tornado==5.1.1
typing-extensions==4.0.0
urllib3==1.26.7
vine==1.3.0
weasyprint==53.4
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.8
xlrd==2.0.1
xlwt==1.3.0
zopfli==0.1.9